Utilities for Age Intervals

This help page documents the utility functions provided for working with individual ages and associated intervals:

breaks_to_interval() takes a specified set of breaks representing the left hand limits of a closed open interval, i.e [x, y), and returns the corresponding interval and upper bounds. The resulting intervals span from the minimum break through to Inf.

cut_ages() provides categorisation of ages based on specified breaks which represent the left-hand interval limits. The resultant groupings will span from the minimum break through to Inf and will always be closed on the left and open on the right. Ages below the minimum break will be returned as NA. As an example, if breaks = c(0, 1, 10, 30) the possible groupings would be [0, 1), [1, 10), [10, 30) and [30, Inf). This is roughly comparable to a call of cut(ages, right = FALSE, breaks = c(limits, Inf)) but with both the resultant interval and the start and end points returned as entries in a list.

split_interval_counts() splits counts of a given age interval in to counts for individual years based on a given weighting. Age intervals are specified by their lower (closed) and upper (open) bounds, i.e. intervals of the form [lower, upper).

aggregate_age_counts() provides aggregation of counts across ages (in years). It is similar to a cut() and tapply() pattern but optimised for speed over flexibility. Groupings are the same as in ages_to_interval() and counts will be provided across all natural numbers grater than the minimum break. Missing values, and those less than the minimum break, are grouped as NA.

reaggregate_interval_counts() is equivalent to, but more efficient than, a call to split_interval_counts() followed by aggregate_age_counts().

Usage

breaks_to_interval(breaks)

cut_ages(ages, breaks)

split_interval_counts(
  lower_bounds,
  upper_bounds,
  counts,
  max_upper = 100L,
  weights = NULL
)

aggregate_age_counts(counts, ages = 0:(length(counts) - 1L), breaks)

reaggregate_interval_counts(
  lower_bounds,
  upper_bounds,
  counts,
  breaks,
  max_upper = 100L,
  weights = NULL
)

Arguments

breaks

[numeric].

1 or more non-negative cut points in increasing (strictly) order.

These correspond to the left hand side of the desired intervals (e.g. the closed side of [x, y).

Double values are coerced to integer prior to categorisation.

ages

[numeric].

Vector of age in years.

Double values are coerced to integer prior to categorisation / aggregation.

For aggregate_age_counts(), these must corresponding to the counts entry and will defaults to 0:(N-1) where N is the number of counts present.

ages >= 200 are not permitted due to the internal implementation.

lower_bounds, upper_bounds

[integerish].

A pair of vectors representing the bounds of the intervals.

lower_bounds must be strictly less than upper_bounds and greater than or equal to zero.

Missing (NA) bounds are not permitted.

Double vectors will be coerced to integer.

counts

[numeric].

Vector of counts to be aggregated.

max_upper

[integerish]

Represents the maximum upper bounds permitted upon splitting the data.

Used to replace Inf upper bounds prior to splitting.

If any upper_bound is greater than max_upper the function will error.

Double vectors will be coerced to integer.

weights

[numeric]

Population weightings to apply for individual years.

If NULL (default) counts will be split evenly based on interval size.

If specified, must be of length max_upper and represent weights in the range 0:(max_upper - 1).

Value

breaks_to_interval() and cut_ages():

A data frame with an ordered factor column (interval), as well as columns corresponding to the explicit bounds (lower_bound and upper_bound).

split_interval_counts():

A data frame with entries age (in years) and count.

aggregate_age_counts() and reaggregate_interval_counts():

A data frame with 4 entries; interval, lower_bound, upper_bound and an associated count.

Examples

cut_ages(ages = 0:9, breaks = c(0L, 3L, 5L, 10L))
#>    interval lower_bound upper_bound
#> 1    [0, 3)           0           3
#> 2    [0, 3)           0           3
#> 3    [0, 3)           0           3
#> 4    [3, 5)           3           5
#> 5    [3, 5)           3           5
#> 6   [5, 10)           5          10
#> 7   [5, 10)           5          10
#> 8   [5, 10)           5          10
#> 9   [5, 10)           5          10
#> 10  [5, 10)           5          10
cut_ages(ages = 0:9, breaks = 5L)
#>    interval lower_bound upper_bound
#> 1      <NA>          NA          NA
#> 2      <NA>          NA          NA
#> 3      <NA>          NA          NA
#> 4      <NA>          NA          NA
#> 5      <NA>          NA          NA
#> 6  [5, Inf)           5         Inf
#> 7  [5, Inf)           5         Inf
#> 8  [5, Inf)           5         Inf
#> 9  [5, Inf)           5         Inf
#> 10 [5, Inf)           5         Inf

split_interval_counts(
    lower_bounds = c(0, 5, 10),
    upper_bounds = c(5, 10, 20),
    counts = c(5, 10, 30)
)
#>    age count
#> 1    0     1
#> 2    1     1
#> 3    2     1
#> 4    3     1
#> 5    4     1
#> 6    5     2
#> 7    6     2
#> 8    7     2
#> 9    8     2
#> 10   9     2
#> 11  10     3
#> 12  11     3
#> 13  12     3
#> 14  13     3
#> 15  14     3
#> 16  15     3
#> 17  16     3
#> 18  17     3
#> 19  18     3
#> 20  19     3

# default ages generated if only counts provided (here ages will be 0:64)
aggregate_age_counts(counts = 1:65, breaks = c(0L, 1L, 5L, 15L, 25L, 45L, 65L))
#>    interval lower_bound upper_bound count
#> 1    [0, 1)           0           1     1
#> 2    [1, 5)           1           5    14
#> 3   [5, 15)           5          15   105
#> 4  [15, 25)          15          25   205
#> 5  [25, 45)          25          45   710
#> 6  [45, 65)          45          65  1110
#> 7 [65, Inf)          65         Inf     0
#> 8      <NA>          NA          NA     0
aggregate_age_counts(counts = 1:65, breaks = 50L)
#>    interval lower_bound upper_bound count
#> 1 [50, Inf)          50         Inf   870
#> 2      <NA>          NA          NA  1275

# NA ages are handled with their own grouping
ages <- 1:65;
ages[1:44] <- NA
aggregate_age_counts(
    counts = 1:65,
    ages = ages,
    breaks = c(0L, 1L, 5L, 15L, 25L, 45L, 65L)
)
#>    interval lower_bound upper_bound count
#> 1    [0, 1)           0           1     0
#> 2    [1, 5)           1           5     0
#> 3   [5, 15)           5          15     0
#> 4  [15, 25)          15          25     0
#> 5  [25, 45)          25          45     0
#> 6  [45, 65)          45          65  1090
#> 7 [65, Inf)          65         Inf    65
#> 8      <NA>          NA          NA   990

reaggregate_interval_counts(
    lower_bounds = c(0, 5, 10),
    upper_bounds = c(5, 10, 20),
    counts = c(5, 10, 30),
    breaks = c(0L, 1L, 5L, 15L, 25L, 45L, 65L)
)
#>    interval lower_bound upper_bound count
#> 1    [0, 1)           0           1     1
#> 2    [1, 5)           1           5     4
#> 3   [5, 15)           5          15    25
#> 4  [15, 25)          15          25    15
#> 5  [25, 45)          25          45     0
#> 6  [45, 65)          45          65     0
#> 7 [65, Inf)          65         Inf     0
#> 8      <NA>          NA          NA     0