Deprecated functions

The functions listed below are deprecated and may be removed in a future package release.

Usage

reaggregate_interval_rates(
  lower_bounds,
  upper_bounds = NULL,
  rates,
  breaks,
  weights = NULL
)

reaggregate_interval_counts(
  lower_bounds,
  upper_bounds,
  counts,
  breaks,
  max_upper = 100L,
  weights = NULL
)

split_interval_counts(
  lower_bounds,
  upper_bounds,
  counts,
  max_upper = 100L,
  weights = NULL
)

aggregate_age_counts(counts, ages = seq_along(counts) - 1L, breaks)

Note

Help pages for each deprecated function can be found via help("<function>-deprecated").

See also